Struct igd::XmlError

pub struct XmlError {
    // some fields omitted
}

XML parsing error.

Consists of a position and a message.

Methods

impl Error

fn new<O>(o: &O, msg: String) -> Error where O: Position

Creates a new error using position information from the provided Position object and a message.

fn msg(&'a self) -> &'a str

Returns a reference to a message which is contained inside this error.