Crate igd

This library allows you to communicate with an IGD enabled device. Use one of the search_gateway functions to obtain a Gateway object. You can then communicate with the device via this object.

Structs

Gateway
This structure represents a gateway found by the search functions.

SearchOptions
Gateway search configuration

Enums

AddAnyPortError
Errors returned by Gateway::add_any_port and Gateway::get_any_address

AddPortError
Errors returned by Gateway::add_port

Error
An error type that emcompasses all possible errors.

GetExternalIpError
Errors returned by Gateway::get_external_ip

PortMappingProtocol
Represents the protocols available for port mapping.

RemovePortError
Errors returned by Gateway::remove_port

RequestError
Errors that can occur when sending the request to the gateway.

SearchError
Errors than can occur while trying to find the gateway.

Functions

search_gateway
Search gateway, using the given SearchOptions.

Type Definitions

Result
A result type where the error is igd::Error.